requirements.txt
argon2-cffi==19.2.0
astroid==2.3.2
bcrypt==3.1.7
beautifulsoup4==4.8.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
django==2.2.7
django-allauth==0.41.0
django-bootstrap4==1.0.1
django-braces==1.13.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.0
django-debug-toolbar==2.2
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
Faker==2.0.4
flake8==3.7.9
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2
mccabe==0.6.1
misaka==2.1.1
numpy==1.18.1
oauthlib==3.1.0
Pillow==6.2.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.8
wrapt==1.11.2
